[MAJOR-14]
* separated the sun, moon, planet, and others in cel__stars.py, cel__moons.py, cel__planets.py, cel__others.py and constants in cel__CONSTANTS.py
* optimized imports in solar_gui.py, and tk_functions.py
* removed celestial_objects.py
